+++ src/extension.ts
@@ -8,10 +8,7 @@
  */
 
 "use strict";
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
-//import * as vscode from "vscode";
-import DotNetAutoAttach from "./dotNetAutoAttach";
+import DotNetWatch from "./dotNetWatch";
 
 /**
  * This method is called when your extension is activated.
@@ -20,8 +17,8 @@ import DotNetAutoAttach from "./dotNetAutoAttach";
  * @export
  * @param {vscode.ExtensionContext} context
  */
-export function activate(/*context: vscode.ExtensionContext*/) {
-  DotNetAutoAttach.Start();
+export function activate() {
+  DotNetWatch.Start();
 }
 
 /**
@@ -30,5 +27,5 @@ export function activate(/*context: vscode.ExtensionContext*/) {
  * @export
  */
 export function deactivate() {
-  DotNetAutoAttach.Stop();
+  DotNetWatch.Stop();
 }