+++ src/services/ui-service.ts
@@ -8,9 +8,7 @@
  */
 
 import { Disposable, QuickPickOptions, Uri, window } from "vscode";
-import { DebugDisconnectedEnum } from "../enums/DebugDisconnectedEnum";
-import DotNetAutoAttachDebugConfiguration from "../interfaces/IDotNetAutoAttachDebugConfiguration";
-import LaunchProfileQuickPickItem from "../models/LaunchProfileQuickPickItem";
+import DotNetWatchDebugConfiguration from "../interfaces/IDotNetWatchDebugConfiguration";
 import ProjectQuickPickItem from "../models/ProjectQuickPickItem";
 
 /**
@@ -61,37 +59,6 @@ export default class UiService implements Disposable {
   }
 
   /**
-   * Opens a Debug Disconnected Information Message.
-   *
-   * @param {string} projectName
-   * @param {number} processId
-   * @returns {(Thenable<string | undefined>)}
-   * @memberof UiService
-   */
-  public DebugDisconnectedInformationMessage(projectName: string, processId: number): Thenable<DebugDisconnectedEnum> {
-    return window
-      .showInformationMessage(
-        `Debug disconnected. Reattach to ${projectName} (${processId}) ?`,
-        "Yes",
-        "No",
-        "Stop watch task"
-      )
-      .then((ret) => {
-        switch (ret) {
-          case "Yes":
-            return DebugDisconnectedEnum.Yes;
-            break;
-          case "Stop watch task":
-            return DebugDisconnectedEnum.Stop;
-            break;
-          default:
-            return DebugDisconnectedEnum.No;
-            break;
-        }
-      });
-  }
-
-  /**
    * Opens a Task already started Information Message.
    *
    * @param {string} projectName
@@ -109,7 +76,7 @@ export default class UiService implements Disposable {
    * @returns {(Thenable<string | undefined>)}
    * @memberof UiService
    */
-  public ProjectDoesNotExistErrorMessage(debugConfig: DotNetAutoAttachDebugConfiguration): Thenable<boolean> {
+  public ProjectDoesNotExistErrorMessage(debugConfig: DotNetWatchDebugConfiguration): Thenable<boolean> {
     return window
       .showErrorMessage(
         `The debug configuration '${debugConfig.name}' within the launch.json references a project that cannot be found or is not unique (${debugConfig.project}).`,
@@ -125,46 +92,7 @@ export default class UiService implements Disposable {
   }
 
   /**
-   * Opens a Project Quick Pick.
-   *
-   * @private
-   * @param {Uri[]} profiles
-   * @returns {(Thenable<LaunchProfileQuickPickItem | undefined>)}
-   * @memberof TaskService
-   */
-  public OpenLaunchProfileQuickPick(
-    projectName: string,
-    profiles: Array<string>
-  ): Thenable<LaunchProfileQuickPickItem | undefined> {
-    const quickPickOptions: QuickPickOptions = {
-      canPickMany: false,
-      placeHolder: `[${projectName}] Which launch profile? (loads environment variables)`,
-      matchOnDescription: true,
-      matchOnDetail: true,
-      ignoreFocusOut: true,
-    };
-    return window.showQuickPick(
-      profiles.map((k) => new LaunchProfileQuickPickItem(k)),
-      quickPickOptions
-    );
-    // const quickPick = window.createQuickPick();
-
-    // quickPick.items = profiles.map((k) => new LaunchProfileQuickPickItem(k));
-    // quickPick.canSelectMany = false;
-    // quickPick.placeholder = `[${projectName}] Which launch profile? (loads environment variables)`;
-    // quickPick.matchOnDescription = true;
-    // quickPick.matchOnDetail = true;
-    // quickPick.show();
-    // quickPick.onDidChangeSelection((selection) => {
-    //   if (selection) quickPick.hide();
-    // });
-    // return new Promise<LaunchProfileQuickPickItem | undefined>((resolve) =>
-    //   quickPick.onDidAccept(() => resolve(quickPick.selectedItems[0]))
-    // );
-  }
-
-  /**
-   * Opens a ProjectDoesNotExist Error Message.
+   * Opens a generic Error Message.
    *
    * @param {string} project
    * @returns {(Thenable<string | undefined>)}